edits

+++ backend/src/services/groupService.ts
@@ -74,22 +74,50 @@ export class GroupService extends BaseService {
   async updateGroup(group: GroupType): Promise<void> {
     return this.performTransaction(async (client) => {
       await groupEntityService.updateGroup(client, group)
+
+      switch (group.type) {
+        case 'standard':
+          const userIds = await userGroupRelEntityService.getRelations(client, group.id)
+          const addedUsers = group.users.filter((user) => !userIds.includes(user.id))
+          await Promise.all(addedUsers.map(async user => {
+            await userGroupRelEntityService.insertRelation(client, user.id, group.id, group.accountId)
+          }))
+          const removedUsers = userIds.filter((userId) => !group.users.map((user) => user.id).includes(userId))
+          await Promise.all(removedUsers.map(async userId => {
+            await userGroupRelEntityService.deleteRelation(client, userId, group.id, group.accountId)
+          }))
+          break
+        case 'intelligent':
+          const ruleRelations = await groupFilterEntityService.selectGroupFilters(client, group.id)
+          const addedRules = group.rules.filter((rule) => !ruleRelations.map((rule) => rule.id).includes(rule.id))
+          await Promise.all(addedRules.map(async rule => {
+            await groupFilterEntityService.insertGroupFilter(client, group.id, rule)
+          }))
+          const removedRules = ruleRelations.filter((rule) => !group.rules.map((rule) => rule.id).includes(rule.id))
+          await Promise.all(removedRules.map(async rule => {
+            await groupFilterEntityService.deleteGroupFilter(client, rule.id)
+          }))
+          break
+      }
     })
   }
 
   async deleteGroup(groupId: string, accountId: string): Promise<void> {
     return this.performTransaction(async (client) => {
-      const userRelations = await userGroupRelEntityService.getRelations(client, groupId)
 
-      await Promise.all(userRelations.map(async userId => {
-        await userGroupRelEntityService.deleteRelation(client, userId, groupId, accountId)
-      }))
+      const userRelations = await userGroupRelEntityService.getRelations(client, groupId)
+      if (userRelations.length !== 0) {
+        await Promise.all(userRelations.map(async userId => {
+          await userGroupRelEntityService.deleteRelation(client, userId, groupId, accountId)
+        }))
+      }
 
       const ruleRelations = await groupFilterEntityService.selectGroupFilters(client, groupId)
-
-      await Promise.all(ruleRelations.map(async rule => {
-        await groupFilterEntityService.deleteGroupFilter(client, rule.id)
-      }))
+      if (ruleRelations.length !== 0) {
+        await Promise.all(ruleRelations.map(async rule => {
+          await groupFilterEntityService.deleteGroupFilter(client, rule.id)
+        }))
+      }
 
       await groupEntityService.deleteGroup(client, groupId) 
     })
